Replace the frontend
The existing frontend fetches the list of repos/tags using the
pagination options available in the registry API. This is essentially
unusable when you have as many repos as us, because navigating to a
specific repo is either:
  a) Incredibly cumbersome when you have to click through all the pages
  b) Incredibly slow if you fetch a large set of repos

This new frontend fetches the list of repos/tags asynchronously and caches
them, meaning that the information is readily available and quick to
navigate once it's cached.

It also gives you more information about the manifests AND captures
events from the registry (i.e who pushed what image when).